ab_system_contract_state::ffi::initialize

Function state_initialize

#[unsafe(no_mangle)]
pub unsafe extern "C" fn state_initialize(
    args: &mut InternalArgs<'_>,
) -> ExitCode

FFI interface into a method, called by the host.

NOTE: Calling this function directly shouldn’t be necessary except maybe in contract’s own tests.

Safety

Caller must ensure the provided pointer corresponds to the expected ABI.