ab_system_contract_state::ffi::read
Struct StateReadArgs

#[repr(C)]
pub struct StateReadArgs {
    pub contract_state_ptr: NonNull<Address>,
    pub state_ptr: NonNull<<VariableBytes<MAX_STATE_SIZE> as IoType>::PointerType>,
    pub state_size: *mut u32,
    pub state_capacity: NonNull<u32>,
    pub ok_result_ptr: NonNull<()>,
    pub ok_result_size: *mut u32,
    pub ok_result_capacity: NonNull<u32>,
}

Data structure containing expected input for external method invocation, eventually calling read() on the other side by the host.

This can be used with Env, though there are helper methods on this provided by extension trait that allow not dealing with this struct directly in simpler cases.

Fields
contract_state_ptr: NonNull<Address>state_ptr: NonNull<<VariableBytes<MAX_STATE_SIZE> as IoType>::PointerType>state_size: *mut u32

Size of the contents state_ptr points to

state_capacity: NonNull<u32>

Capacity of the allocated memory state_ptr points to

ok_result_ptr: NonNull<()>ok_result_size: *mut u32

Size of the contents ok_result_ptr points to

ok_result_capacity: NonNull<u32>

Capacity of the allocated memory ok_result_ptr points to